Extract `{RustInstaller,Distcheck,CollectLicenseMetadata}` into their own `dist_tool_tests` module
